+++ apps/docs/src/content/03-patterns/01-patterns/anlegeprozess/examples/install-extension.tsx
@@ -26,7 +26,7 @@ import {
 } from "@mittwald/flow-react-components";
 
 <ModalTrigger>
-  <Button color="accent">Beispiel öffnen</Button>
+  <Button color="success">Beispiel öffnen</Button>
   <Modal offCanvas>
     <Heading>Passwortschutz installieren</Heading>
     <Content>
@@ -54,7 +54,7 @@ import {
               Installation fortzufahren, lege bitte einen
               Vertragspartner an.
             </Text>
-            <Button color="accent">
+            <Button color="success">
               Vertragspartner anlegen
             </Button>
           </Content>