+++ apps/docs/src/content/04-components/actions/action-group/examples/link.tsx
@@ -6,5 +6,5 @@ import {
 
 <ActionGroup>
   <Link slot="abort">Passwort vergessen?</Link>
-  <Button color="accent">Anmelden</Button>
+  <Button color="success">Anmelden</Button>
 </ActionGroup>;