+++ apps/docs/src/content/04-components/actions/action-group/examples/slots.tsx
@@ -4,7 +4,7 @@ import {
 } from "@mittwald/flow-react-components";
 
 <ActionGroup>
-  <Button color="accent">E-Mail-Addresse anlegen</Button>
+  <Button color="success">E-Mail-Addresse anlegen</Button>
   <Button color="secondary" variant="soft" slot="secondary">
     Speichern und weitere anlegen
   </Button>