+++ apps/docs/src/content/04-components/actions/button/examples/hierarchy-soft-do.tsx
@@ -1,8 +1,0 @@
-import { Button } from "@mittwald/flow-react-components";
-
-<>
-  <Button variant="soft" color="secondary">
-    Zurück
-  </Button>
-  <Button>Weiter</Button>
-</>;