+++ apps/docs/src/content/04-components/actions/button/examples/sizes-dont.tsx
@@ -1,7 +1,0 @@
-import { Button } from "@mittwald/flow-react-components";
-
-<>
-  <Button>Weiter</Button>
-  <Button size="s">Weiter</Button>
-  <Button size="s">Weiter</Button>
-</>;