+++ apps/docs/src/content/04-components/content/label/examples/do.tsx
@@ -1,8 +1,0 @@
-import { Label } from "@mittwald/flow-react-components";
-
-<>
-  <Label>Vorname</Label>
-  <Label>E-Mail</Label>
-  <Label>Verknüpfte App</Label>
-  <Label>Datenbankbenutzer</Label>
-</>;