+++ packages/components/src/components/Combine/Combine.module.d.scss.ts
@@ -1,0 +1,11 @@
+declare const classNames: {
+  readonly combine: "combine";
+  readonly avatar: "avatar";
+  readonly text: "text";
+  readonly "flow--form-field": "flow--form-field";
+  readonly button: "button";
+  readonly contextualHelpTriggerButton: "contextualHelpTriggerButton";
+  readonly copyButton: "copyButton";
+  readonly icon: "icon";
+};
+export default classNames;