+++ packages/remote-react-components/src/tests/visual/Link.browser.test.tsx
@@ -88,7 +88,7 @@ test.each(testEnvironments)(
           <Button>Small & Download</Button>
         </Link>
         <Link>
-          <Button color="accent">Button Color</Button>
+          <Button color="success">Button Color</Button>
         </Link>
         <Link color="dark">
           <Button>Link Color</Button>